Time: 6 ms (24.51%), Space: 47 MB (81.23%) - LeetHub

+++ 637-average-of-levels-in-binary-tree/637-average-of-levels-in-binary-tree.java
@@ -32,7 +32,7 @@ class Solution {
                 
                 ans += q.poll().val;
             }
-            res.add(ans/size);
+            res.add((double)ans/size);
         }
